Reworked the user-settable build variables
Added CPPFLAGS, which may be used in commands where CFLAGS isn't.

PKG_CONFIG_PATH is left empty for the user to override,
and the value passed to sub-builds is composed from that
and the build/installed location.

PKG_CONFIG_PATH still has to be passed in the environment of
gobject-introspection sub-make because of
https://bugzilla.gnome.org/show_bug.cgi?id=757128